# openapi/minimovie-api.yaml
openapi: 3.0.3
info:
  title: MiniMovie API
  description: API for retrieving intuitive movie, TV series, and person information
  version: 1.0.0
servers:
  - url: https://api.minimovie.info
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /auth/apple/notifications:
    post:
      summary: Apple server-to-server notifications
      description: >-
        Receives Apple Sign in with Apple server-to-server event notifications.
        Validates the JWS payload signature against Apple's JWKS, deduplicates
        by JTI, and dispatches events (consent-revoked, account-delete) to
        delete the associated user account.
      operationId: appleNotifications
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - payload
              properties:
                payload:
                  type: string
                  description: JWS-signed notification from Apple
      responses:
        "200":
          description: Notification processed
        "400":
          description: Invalid or stale payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid signature or claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/me/stats:
    get:
      summary: Get user stats
      description: Returns computed statistics for the authenticated user's watch history
      operationId: getStats
      security:
        - sessionCookie: []
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Guest API key used for public catalog routes and token exchange
    sessionCookie:
      type: apiKey
      in: cookie
      name: mm_session
      description: Session cookie set after OAuth login (production name is __Host-mm_session)
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: HTTP status text
        message:
          type: string
          description: Detailed error message
...
    StatsResult:
      type: object
      properties:
        totalMoviesWatched:
          type: integer
        totalSeriesCompleted:
          type: integer
        totalEpisodesWatched:
          type: integer
        estimatedMinutesWatched:
          type: integer
        genreBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/GenreStat"
        monthlyActivity:
          type: array
          items:
            $ref: "#/components/schemas/MonthActivity"
        currentStreak:
          type: integer
          description: Current consecutive days with watch activity
        longestStreak:
          type: integer
          description: Longest consecutive days with watch activity
    GenreStat:
      type: object
      properties:
        genre:
          type: string
        count:
          type: integer
    MonthActivity:
      type: object
      properties:
        month:
          type: string
          description: Year-month in YYYY-MM format
          example: "2026-04"
        moviesWatched:
          type: integer
        episodesWatched:
          type: integer
...